+++ spec/syncer_spec.rb
@@ -7,10 +7,8 @@ describe NetworkResiliency::Syncer do
     allow(NetworkResiliency::Syncer).to receive(:start).and_call_original
   end
 
-  let(:redis) { Redis.new }
-
   def start
-    described_class.start(redis)
+    described_class.start
   end
 
   describe ".start" do
@@ -18,16 +16,36 @@ describe NetworkResiliency::Syncer do
       expect(start).to be_a(Thread)
     end
 
-    it "can be called many times without error" do
-      3.times { start }
+    it "is idempotent" do
+      threads = 3.times.map { start }
+      expect(threads.uniq.count).to eq 1
     end
 
-    it "will stop previous workers so only one is running at a time" do
-      first_worker = start
-      second_worker = start
+    it "is shared between threads" do
+      t = Thread.new { start }
 
-      expect(first_worker).not_to be_alive
-      expect(second_worker).to be_alive
+      expect(t.value).to eq start
+    end
+
+    context "when worker is stopped" do
+      let(:worker) do
+        worker = start
+        worker.kill
+        worker.join
+      end
+
+      it "starts a new worker" do
+        expect(worker).not_to be_alive
+        expect(start).not_to eq worker
+      end
+    end
+
+    context "when redis is not configured" do
+      before { NetworkResiliency.redis = nil }
+
+      it "does not start a worker" do
+        expect(start).to be_nil
+      end
     end
   end
 
@@ -59,6 +77,35 @@ describe NetworkResiliency::Syncer do
       described_class.stop
 
       is_expected.to be false
+    end
+  end
+
+  describe "#sync" do
+    subject(:sync) { worker.send(:sync) }
+
+    let(:worker) do
+      worker = start
+      worker.kill
+      worker.join
+    end
+
+    context "when worker encounters an error", :safely do
+      before do
+        allow(NetworkResiliency::StatsEngine).to receive(:sync).and_raise
+      end
+
+      it "logs the error" do
+        expect(NetworkResiliency).to receive(:warn)
+
+        sync
+      end
+    end
+
+    it "sleeps between syncs" do
+      expect(NetworkResiliency::StatsEngine).to receive(:sync)
+      expect(worker).to receive(:sleep).and_raise(Interrupt)
+
+      sync
     end
   end
 end